# docs/swagger.yaml
basePath: /
definitions:
  dto.AuthResponse:
    properties:
      token:
        type: string
      user:
        $ref: '#/definitions/dto.UserResponse'
    type: object
  dto.ErrorResponse:
    properties:
      error:
        type: string
      message:
        type: string
    type: object
...
  dto.UserResponse:
    properties:
      activities:
        type: string
      allergic_drugs:
        type: string
      allergic_food:
        type: string
      avatar_url:
        type: string
      birth_date:
        type: string
      chronic_disease:
        type: string
      created_at:
        type: string
      display_name:
        type: string
      email:
        type: string
      emergency_contact:
        type: string
      food_categories:
        type: string
      food_preferences:
        type: string
      id:
        type: string
      phone:
        type: string
      role:
        type: string
      updated_at:
        type: string
      username:
        type: string
    type: object
...
host: localhost:8080
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: Go2gether Backend API for travel companion matching
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: http://swagger.io/terms/
  title: Go2gether Backend API
  version: "1.0"
paths:
...
  /api/auth/google/callback:
    get:
      consumes:
      - application/json
      description: Handle Google OAuth callback with authorization code
      parameters:
      - description: Authorization code from Google
        in: query
        name: code
        required: true
        type: string
      - description: State parameter for CSRF protection
        in: query
        name: state
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Login successful
          schema:
            $ref: '#/definitions/dto.AuthResponse'
        "400":
          description: Invalid request data
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "401":
          description: Invalid authorization code
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      summary: Google OAuth callback
      tags:
      - authentication
...
schemes:
- http
- https
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"